Prepare for main merge with new variations_router
